+++ Gymify.Model/ResponseObjects/MembershipPackageAnalyticsResponse.cs
@@ -5,6 +5,6 @@ namespace Gymify.Model.ResponseObjects
         public int MembershipId { get; set; }
         public string MembershipName { get; set; } = string.Empty;
         public int PurchaseCount { get; set; }
-        public decimal TotalIncome { get; set; }
+        public int TotalIncome { get; set; }
     }
 }
