+++ Gymify.Model/ResponseObjects/MembershipResponse.cs
@@ -7,9 +7,9 @@ namespace Gymify.Model.ResponseObjects
         public int Id { get; set; }
         public string Name { get; set; }
 
-        public double MonthlyPrice { get; set; }
+        public int MonthlyPrice { get; set; }
 
-        public double YearPrice { get; set; }
+        public int YearPrice { get; set; }
 
         public DateTime CreatedAt { get; set; }
     }
